CDI ContextualLifecycle for JavaEE
